Phase 5: split algorithms/base.py into ABC-per-module layout
Mechanical split of the 1238-line algorithms/base.py so base.py holds only
the Algorithm ABC, matching the package idiom where base.py is the ABC module
(solvers/base.py, discretization/base.py, integrations/base.py). Pure code
movement plus import updates; no logic changes.

New modules (acyclic import order hyperparams -> state -> history):
- hyperparams.py: HyperParams + its module-level base registration.
- state.py: AlgorithmState (+ _FIELDS, export registration), AdaptiveStateCode
  (+ names map, adaptive_state_code_to_str), CandidateIterate.
- history.py: AlgorithmHistory, DiscretizationResult.
- penalty.py: calculate_cost_from_state / calculate_nonlinear_penalty as
  module functions (dropped from AutotuningBase; the latter now calls the
  former directly).
- autotuner/base.py: AutotuningBase + LamCostRelaxHyper.

base.py keeps the Algorithm ABC, importing state/history/autotuner.base/weights
under TYPE_CHECKING with quoted annotations.

The jax.export serialized names for AlgorithmState and the HyperParams base are
kept as the historical "openscvx.algorithms.base.*" strings (commented as stable
serialization IDs, not module paths) so cached solve_batched artifacts still
deserialize. Confirmed load_or_export_solve_batched hard-errors rather than
re-exports on an unknown serialized_name, and noted that in caching.py.

Call sites (autotuner penalty calls, ~22 test calls, internal imports across
scvx/autotuner/optimization_results and four test modules) repointed to the new
homes; deleted the two now-dead hasattr asserts on the moved penalty math.
